refactor: made utility function retrieveAllApexTestItemsFromSidebar() in apexUtils.ts
