+++ test/specs/debugApexTests.e2e.ts
@@ -119,25 +119,7 @@ describe('Debug Apex Tests', async () => {
     const apexTestsSection = await sidebarView.getSection('APEX TESTS');
     expect(apexTestsSection.elem).toBePresent();
 
-    let apexTestsItems = (await apexTestsSection.getVisibleItems()) as TreeItem[];
-    await browser.keys(['Escape']);
-
-    // If the Apex tests did not show up, click the refresh button on the top right corner of the Test sidebar
-    for (let x = 0; x < 3; x++) {
-      if (apexTestsItems.length === 1) {
-        await apexTestsSection.elem.click();
-        const refreshAction = await apexTestsSection.getAction('Refresh Tests');
-        await refreshAction!.elem.click();
-        utilities.pause(10);
-        apexTestsItems = (await apexTestsSection.getVisibleItems()) as TreeItem[];
-      }
-      else if (apexTestsItems.length === 4) {
-        break;
-      }
-      else {
-        // do nothing
-      }
-    }
+    let apexTestsItems = await utilities.retrieveAllApexTestItemsFromSidebar(4, apexTestsSection);
 
     // Make sure all the tests are present in the sidebar
     expect(apexTestsItems.length).toBe(4);
